requirements.txt
ansicon==1.89.0
asgiref==3.5.2
asttokens==2.0.7
autopep8==1.7.0
backcall==0.2.0
blessed==1.19.1
codefind==0.1.3
colorama==0.4.5
decorator==5.1.1
distlib==0.3.5
Django==4.1
django-cors-headers==3.13.0
django-taggit==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
executing==0.10.0
filelock==3.8.0
fontawesomefree==6.1.2
ipython==8.4.0
jedi==0.18.1
jinxed==1.2.0
jurigged==0.5.2
matplotlib-inline==0.1.3
ovld==0.3.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.12.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
toml==0.10.2
traitlets==5.3.0
tzdata==2022.1
virtualenv==20.16.3
watchdog==2.1.9
wcwidth==0.2.5
ansicon==1.89.0
asgiref==3.5.2
asttokens==2.0.7
autopep8==1.7.0
backcall==0.2.0
blessed==1.19.1
codefind==0.1.3
colorama==0.4.5
decorator==5.1.1
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-heroku==0.3.1
django-taggit==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
executing==0.10.0
filelock==3.8.0
fontawesomefree==6.1.2
ipython==8.4.0
jedi==0.18.1
jinxed==1.2.0
jurigged==0.5.2
matplotlib-inline==0.1.3
ovld==0.3.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.12.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
toml==0.10.2
traitlets==5.3.0
tzdata==2022.1
virtualenv==20.16.3
watchdog==2.1.9
wcwidth==0.2.5
whitenoise==6.2.0
ansicon==1.89.0
asgiref==3.5.2
asttokens==2.0.7
autopep8==1.7.0
backcall==0.2.0
blessed==1.19.1
codefind==0.1.3
colorama==0.4.5
decorator==5.1.1
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-heroku==0.3.1
django-taggit==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
executing==0.10.0
filelock==3.8.0
fontawesomefree==6.1.2
gunicorn==20.1.0
ipython==8.4.0
jedi==0.18.1
jinxed==1.2.0
jurigged==0.5.2
matplotlib-inline==0.1.3
ovld==0.3.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.12.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
toml==0.10.2
traitlets==5.3.0
tzdata==2022.1
virtualenv==20.16.3
watchdog==2.1.9
wcwidth==0.2.5
whitenoise==6.2.0
